Refactor activity list items to separate items with stats from items without
