+++ src/features/campaigns/components/ActivityList/index.tsx
@@ -3,11 +3,11 @@ import Fuse from 'fuse.js';
 import { useMemo } from 'react';
 import { Box, Card, Divider, Typography } from '@mui/material';
 
-import CallAssignmentListItem from './CallAssignmentListItem';
-import EventListItem from './EventListItem';
+import CallAssignmentListItem from './items/CallAssignmentListItem';
+import EventListItem from './items/EventListItem';
 import messageIds from 'features/campaigns/l10n/messageIds';
-import SurveyListItem from './SurveyListItem';
-import TaskListItem from './TaskListItem';
+import SurveyListItem from './items/SurveyListItem';
+import TaskListItem from './items/TaskListItem';
 import { useMessages } from 'core/i18n';
 import {
   ACTIVITIES,
